requirements.txt
requests==2.32.3
pyyaml==6.0.2
pydantic==2.9.2
language-tool-python==2.8.1
sentence-transformers==3.0.1
torch==2.3.1
numpy==2.0.2
